Refactor link extraction in tests
