+++ tests/api/subscriptions.rs
@@ -135,24 +135,8 @@ async fn subscribe_sends_a_confirmation_email_with_a_link() {
     // Assert
     // Get the first intercepted request
     let email_request = &app.email_server.received_requests().await.unwrap()[0];
-
-    // Parse the body as JSON
-    let body: serde_json::Value = email_request.body_json().unwrap();
-
-    // Extract the link from one of the request fields.
-    let get_link = |s: &str| {
-        let links: Vec<_> = linkify::LinkFinder::new()
-            .links(s)
-            .filter(|l| *l.kind() == linkify::LinkKind::Url)
-            .collect();
-
-        assert_eq!(links.len(), 1);
-        links[0].as_str().to_owned()
-    };
-
-    let html_link = get_link(body["HtmlBody"].as_str().unwrap());
-    let text_link = get_link(body["TextBody"].as_str().unwrap());
+    let confirmation_links = app.get_confirmation_links(email_request);
 
     // The two links should be identical
-    assert_eq!(html_link, text_link);
+    assert_eq!(confirmation_links.html, confirmation_links.plain_text);
 }
