+++ tests/api/subscriptions_confirm.rs
@@ -1,4 +1,3 @@
-use reqwest::Url;
 use wiremock::matchers::{method, path};
 use wiremock::{Mock, ResponseTemplate};
 
@@ -32,29 +31,10 @@ async fn the_link_returned_by_subscribe_returns_a_200_if_called() {
 
     app.post_subscriptions(body.into()).await;
     let email_request = &app.email_server.received_requests().await.unwrap()[0];
-    let body: serde_json::Value = email_request.body_json().unwrap();
-
-    // Extract the link from one of the request fields.
-    let get_link = |s: &str| {
-        let links: Vec<_> = linkify::LinkFinder::new()
-            .links(s)
-            .filter(|l| *l.kind() == linkify::LinkKind::Url)
-            .collect();
-
-        assert_eq!(links.len(), 1);
-        links[0].as_str().to_owned()
-    };
-    let raw_confirmation_link = &get_link(body["HtmlBody"].as_str().unwrap());
-    let mut confirmation_link = Url::parse(raw_confirmation_link).unwrap();
-
-    // Let's make sure we don't call random APIs on the web
-    assert_eq!(confirmation_link.host_str().unwrap(), "127.0.0.1");
-
-    // Let's rewrite the URL to include the port
-    confirmation_link.set_port(Some(app.port)).unwrap();
+    let confirmation_links = app.get_confirmation_links(email_request);
 
     // Act
-    let response = reqwest::get(confirmation_link).await.unwrap();
+    let response = reqwest::get(confirmation_links.html).await.unwrap();
 
     // Assert
     assert_eq!(response.status().as_u16(), 200);
